# backend/docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "智时助手 (Cortex Assistant) API",
    "description": "\n    ## AI-powered intelligent assistant for Chinese knowledge workers\n\n    This API provides comprehensive functionality for:\n    \n    * **🤖 AI Provider Management** - Configure and manage multiple AI models (text/image)\n    * **📋 Task Management** - AI-powered task generation with Eisenhower Matrix\n    * **💬 Real-time Chat** - WebSocket-based chat with AI models and thinking visualization\n    * **👤 User Profiling** - Big Five personality model and work relationship management\n    * **🖼️ OCR Integration** - Extract text from images using EasyOCR and AI vision models\n    * **🔐 JWT Authentication** - Secure user authentication and authorization\n\n    ### Key Features:\n    - **Multi-modal AI Support**: Text generation, image OCR, and vision-language models\n    - **Real-time Streaming**: WebSocket-based chat with background task management\n    - **Personality-based Customization**: Big Five model for personalized AI interactions\n    - **Eisenhower Matrix**: Task prioritization with urgency/importance classification\n    - **Chinese Language Optimized**: Designed specifically for Chinese knowledge workers\n    ",
    "contact": {
      "name": "Cortex Assistant Team",
      "url": "https://github.com/your-org/helpful-assistant",
      "email": "support@cortex-assistant.com"
    },
    "license": {
      "name": "MIT License",
...
    },
    "version": "2.0.0"
  },
  "servers": [
...
  "paths": {
...
    "/api/ai-providers": {
      "get": {
        "tags": [
          "AI Providers"
        ],
        "summary": "Get Ai Providers",
        "description": "Get all AI providers for the current user",
        "operationId": "get_ai_providers_api_ai_providers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AIProviderResponse"
                  },
                  "type": "array",
                  "title": "Response Get Ai Providers Api Ai Providers Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "AI Providers"
        ],
        "summary": "Create Ai Provider",
        "description": "Create a new AI provider configuration",
        "operationId": "create_ai_provider_api_ai_providers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIProviderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIProviderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/ai-providers/{provider_id}": {
      "put": {
        "tags": [
          "AI Providers"
        ],
        "summary": "Update Ai Provider",
        "description": "Update an AI provider configuration",
        "operationId": "update_ai_provider_api_ai_providers__provider_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Provider Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIProviderUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIProviderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "AI Providers"
        ],
        "summary": "Delete Ai Provider",
        "description": "Delete an AI provider configuration",
        "operationId": "delete_ai_provider_api_ai_providers__provider_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Provider Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai-providers/{provider_id}/test": {
      "post": {
        "tags": [
          "AI Providers"
        ],
        "summary": "Test Ai Provider",
        "description": "Test an AI provider configuration",
        "operationId": "test_ai_provider_api_ai_providers__provider_id__test_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Provider Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai-providers/active/{category}": {
      "get": {
        "tags": [
          "AI Providers"
        ],
        "summary": "Get Active Provider By Category",
        "description": "Get the active AI provider for a specific category",
        "operationId": "get_active_provider_by_category_api_ai_providers_active__category__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIProviderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai-providers/text-models": {
      "get": {
        "tags": [
          "AI Providers"
        ],
        "summary": "Get Available Text Models",
        "description": "Get all ACTIVE text models for chat selection",
        "operationId": "get_available_text_models_api_ai_providers_text_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AIProviderResponse"
                  },
                  "type": "array",
                  "title": "Response Get Available Text Models Api Ai Providers Text Models Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/ai-providers/active": {
      "get": {
        "tags": [
          "AI Providers"
        ],
        "summary": "Get Active Provider",
        "description": "Get the active text AI provider for backward compatibility",
        "operationId": "get_active_provider_api_ai_providers_active_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIProviderResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AIProviderCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "provider_type": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Provider Type"
          },
          "category": {
            "type": "string",
            "pattern": "^(text|image)$",
            "title": "Category",
            "default": "text"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          }
        },
        "type": "object",
        "required": [
          "name",
          "provider_type"
        ],
        "title": "AIProviderCreate"
      },
      "AIProviderResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "provider_type": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Provider Type"
          },
          "category": {
            "type": "string",
            "pattern": "^(text|image)$",
            "title": "Category",
            "default": "text"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "last_tested": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Tested"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "provider_type",
          "id",
          "is_active",
          "created_at"
        ],
        "title": "AIProviderResponse"
      },
      "AIProviderUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(text|image)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "AIProviderUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}